fix: refactored tabs JS controllers
