+++ application/app/helpers/tabs_helper.rb
@@ -5,15 +5,11 @@ module TabsHelper
     "tab-label-#{object_with_id.id}"
   end
 
-  # Returns just the anchor ID, e.g. "tab-123"
-  def tab_anchor_for(object_with_id)
-    return '' if object_with_id.nil?
-
+  def tab_id_for(object_with_id)
     "tab-#{object_with_id.id}"
   end
 
-  # Returns a full href string, e.g. "#tab-123"
-  def tab_href_for(object_with_id)
-    "##{tab_anchor_for(object_with_id)}"
+  def tab_anchor_for(object_with_id)
+    "tab-link-#{object_with_id.id}"
   end
 end
