+++ application/app/views/projects/_project_summary.html.erb
@@ -30,16 +30,16 @@
     <%= render 'shared/tabs_scrollable', id: "tabs-#{project.id}", container_classes: 'me-2' do %>
       <li class="nav-item" role="presentation">
         <a class="nav-link text-secondary active" id="<%= tab_label_for(project) %>" data-bs-toggle="tab"
-           href="<%= tab_href_for(project) %>" role="tab"
-           aria-controls="<%= tab_anchor_for(project) %>" aria-selected="true">
+           href="#<%= tab_anchor_for(project) %>" role="tab" data-bs-target="#<%= tab_id_for(project) %>"
+           aria-controls="<%= tab_id_for(project) %>" aria-selected="true">
           <i class="bi bi-download me-1" aria-hidden="true"></i><%= t('.tab_download_label') %>
         </a>
       </li>
       <% project.upload_bundles.each do |bundle| %>
         <li class="nav-item" role="presentation">
           <a class="nav-link text-secondary" id="<%= tab_label_for(bundle) %>" data-bs-toggle="tab"
-             href="<%= tab_href_for(bundle) %>" role="tab"
-             aria-controls="<%= tab_anchor_for(bundle) %>" aria-selected="false">
+             href="#<%= tab_anchor_for(bundle) %>" role="tab" data-bs-target="#<%= tab_id_for(bundle) %>"
+             aria-controls="<%= tab_id_for(bundle) %>" aria-selected="false">
             <i class="bi bi-upload me-1" aria-hidden="true"></i><%= bundle.name %>
           </a>
         </li>
@@ -49,12 +49,12 @@
 
     <!-- Tab Content -->
     <div class="tab-content">
-      <div class="tab-pane fade show active" id="<%= tab_anchor_for(project) %>" role="tabpanel"
+      <div class="tab-pane fade show active" id="<%= tab_id_for(project) %>" role="tabpanel"
            aria-labelledby="<%= tab_label_for(project) %>">
         <%= render partial: '/projects/project_summary_downloads', locals: { project: project } %>
       </div>
       <% project.upload_bundles.each do |bundle| %>
-        <div class="tab-pane fade" id="<%= tab_anchor_for(bundle) %>" role="tabpanel"
+        <div class="tab-pane fade" id="<%= tab_id_for(bundle) %>" role="tabpanel"
              aria-labelledby="<%= tab_label_for(bundle) %>">
           <%= render partial: '/projects/project_summary_bundle', locals: { project: project, bundle: bundle } %>
         </div>
