+++ application/app/views/projects/show/_download_files.html.erb
@@ -1,5 +1,5 @@
 <% files = project.download_files %>
-<div id="<%= tab_anchor_for(project) %>" class="tab-pane fade show active" role="tabpanel" aria-labelledby="<%= tab_label_for(project) %>">
+<div id="<%= tab_id_for(project) %>" class="tab-pane fade show active" role="tabpanel" aria-labelledby="<%= tab_label_for(project) %>">
   <%= render partial: '/projects/show/download_actions', locals: { project: project } %>
 
   <ul class="container files rounded border">
@@ -49,7 +49,7 @@
             modal_subtitle: file.filename,
             modal_content: t('.modal_delete_confirmation_content'),
           } do %>
-            <%= hidden_field_tag :anchor, 'tab-download-files' %>
+            <%= hidden_field_tag :anchor, tab_anchor_for(project) %>
           <% end %>
 
         </div>
