+++ application/app/views/projects/show/_project_tabs.html.erb
@@ -1,8 +1,8 @@
 <%= render 'shared/tabs_scrollable', id: 'project-tabs', container_classes: 'mb-3' do %>
   <li class="nav-item" role="presentation">
     <a class="nav-link text-secondary active" id="<%= tab_label_for(project) %>" data-bs-toggle="tab"
-       href="<%= tab_href_for(project) %>" role="tab"
-       aria-controls="<%= tab_anchor_for(project) %>" aria-selected="true">
+       href="#<%= tab_anchor_for(project) %>" role="tab" data-bs-target="#<%= tab_id_for(project) %>"
+       aria-controls="<%= tab_id_for(project) %>" aria-selected="true">
       <i class="bi bi-download me-1" aria-hidden="true"></i><%= t('projects.show.project_tabs.tab_download_files_text') %>
     </a>
   </li>
@@ -10,8 +10,8 @@
   <% project.upload_bundles.each do |upload_bundle| %>
     <li class="nav-item" role="presentation">
       <a class="nav-link text-secondary" id="<%= tab_label_for(upload_bundle) %>" data-bs-toggle="tab"
-         href="<%= tab_href_for(upload_bundle) %>" role="tab"
-         aria-controls="<%= tab_anchor_for(upload_bundle) %>" aria-selected="false">
+         href="#<%= tab_anchor_for(upload_bundle) %>" role="tab" data-bs-target="#<%= tab_id_for(upload_bundle) %>"
+         aria-controls="<%= tab_id_for(upload_bundle) %>" aria-selected="false">
         <i class="bi bi-upload me-1" aria-hidden="true"></i><%= upload_bundle.name %>
       </a>
     </li>
