+++ application/test/helpers/tabs_helper_test.rb
@@ -11,12 +11,12 @@ class TabsHelperTest < ActionView::TestCase
 
   test 'tab_anchor_for should return correct anchor ID for Project' do
     project = create_project
-    assert_equal "tab-#{project.id}", tab_anchor_for(project)
+    assert_equal "tab-link-#{project.id}", tab_anchor_for(project)
   end
 
-  test 'tab_href_for should return correct href for Project' do
+  test 'tab_id_for should return correct href for Project' do
     project = create_project
-    assert_equal "#tab-#{project.id}", tab_href_for(project)
+    assert_equal "tab-#{project.id}", tab_id_for(project)
   end
 
   test 'tab_label_for should return correct label ID for UploadBatch' do
@@ -26,11 +26,11 @@ class TabsHelperTest < ActionView::TestCase
 
   test 'tab_anchor_for should return correct anchor ID for UploadBatch' do
     batch = create_upload_bundle(create_project)
-    assert_equal "tab-#{batch.id}", tab_anchor_for(batch)
+    assert_equal "tab-link-#{batch.id}", tab_anchor_for(batch)
   end
 
-  test 'tab_href_for should return correct href for UploadBatch' do
+  test 'tab_id_for should return correct href for UploadBatch' do
     batch = create_upload_bundle(create_project)
-    assert_equal "#tab-#{batch.id}", tab_href_for(batch)
+    assert_equal "tab-#{batch.id}", tab_id_for(batch)
   end
 end
